@@ -3,16 +3,8 @@ import 'package:quiver/time.dart';
 import 'constants.dart';
 
 class MonthView extends StatelessWidget {
-  const MonthView(
-    this.year,
-    this.month,
-    this.events,
-    {
-      Key key,
-      this.onTapHandler,
-      this.theme
-    }
-  );
+  const MonthView(this.year, this.month, this.events,
+      {Key? key, required this.onTapHandler, required this.theme});
 
   final int year;
   final int month;
@@ -20,32 +12,25 @@ class MonthView extends StatelessWidget {
   final Function onTapHandler;
   final ThemeData theme;
 
-  Widget dayMarker(int day, bool hasEvent) => (
-    Expanded(
-      child: GestureDetector(
-        onTap: hasEvent
-          ? () => onTapHandler(day)
-          : () => onTapHandler(0),
-        child: Container(
-          padding: EdgeInsets.all(8.0),
-          decoration: BoxDecoration(
-            color: hasEvent
-              ? theme.accentColor
-              : theme.primaryColor,
-            shape: BoxShape.circle,
-          ),
-          child: Center(
-            child: Text(
+  Widget dayMarker(int day, bool hasEvent) => (Expanded(
+        child: GestureDetector(
+          onTap: hasEvent ? () => onTapHandler(day) : () => onTapHandler(0),
+          child: Container(
+            padding: EdgeInsets.all(8.0),
+            decoration: BoxDecoration(
+              color: hasEvent ? theme.accentColor : theme.primaryColor,
+              shape: BoxShape.circle,
+            ),
+            child: Center(
+                child: Text(
               day.toString(),
               style: hasEvent
-                ? theme.accentTextTheme.body1
-                : theme.textTheme.body1,
-            )
+                  ? theme.accentTextTheme.body1
+                  : theme.textTheme.body1,
+            )),
           ),
         ),
-      ),
-    )
-  );
+      ));
 
   Widget weekRow(int startDay, int lastDay) {
     List<Widget> days = [];
@@ -54,16 +39,17 @@ class MonthView extends StatelessWidget {
         final hasEvent = events[i] != null;
         days.add(dayMarker(i, hasEvent));
       } else {
-        days.add(Expanded(child: Container(),));
+        days.add(Expanded(
+          child: Container(),
+        ));
       }
     }
 
     return Row(children: days);
   }
 
-  Widget weekdayItem(String text) => (
-    Expanded(
-      child: Container(
+  Widget weekdayItem(String text) => (Expanded(
+          child: Container(
         padding: EdgeInsets.all(2.0),
         child: Center(
           child: Text(
@@ -71,14 +57,10 @@ class MonthView extends StatelessWidget {
             style: theme.textTheme.title,
           ),
         ),
-      )
-    )
-  );
+      )));
 
   Widget weekdayRow() {
-    final dayText = ShortDays.map((day) => (
-      weekdayItem(day)
-    )).toList();
+    final dayText = ShortDays.map((day) => (weekdayItem(day))).toList();
 
     return Row(
       children: dayText,
@@ -99,7 +81,10 @@ class MonthView extends StatelessWidget {
         padding: const EdgeInsets.only(top: 4.0),
         child: weekdayRow(),
       ),
-      Divider(color: theme.dividerColor, height: 8.0,),
+      Divider(
+        color: theme.dividerColor,
+        height: 8.0,
+      ),
     ];
 
     while (weekStart <= lastDayOfMonth) {
